requirements.txt
flask
gunicorn
mysql-connector-python
nltk
numpy
scikit-learn
